replaced ${framework} with %{framework(oXbytei)} editor variable
This tries to fix a bug in oXygen's framework extension scripts.  It
seems that there are some things broken. The editor variable expansion
is one thing to consider. ${framework} is expanded to the current
framework and may fail, when it is intended to expand to a base or
intermediate framework. Using ${framework(oXbytei)} enforces to look
up in the right framework.