keep named entities
The XML parser expands the entities when reading the file as XSLT
source. So we have to replace the expanded entities with entities
again in the output. We can only achieve this with the help of a
mapping of entities (identifiers) and paths. We use a project registry
for this purpose.