split action for apparatus markup
For applying the XSLTOperation on an anchor only, i.e. getting an
apparatus entry at the end of a variant passage, we have to split
SurroundWithAnchorsXSLTOperation into SurroundWithAnchorsOperation and
XSLTOperation. The new action 'anchors.apparatus' is part of this
splitting. 'markup.apparatus' is only an XSLTOperation, now.

The XSL now has to do much lesser than before.